+++ R/parse-data.R
@@ -32,7 +32,8 @@ parse_df <- function(data, add_names = FALSE) {
     FUN = function(x) {
       if (inherits(x, "Date")) {
         # as.numeric(x) * 86400000
-        format(x)
+        # format(x)
+        js_date(x)
       } else if (inherits(x, "POSIXt")) {
         as.numeric(x) * 1000
       } else if (inherits(x, "factor")) {
@@ -62,3 +63,49 @@ parse_df <- function(data, add_names = FALSE) {
   )
   return(ll)
 }
+
+
+
+#' @importFrom htmlwidgets JS
+js_date <- function(x) {
+  lapply(sprintf("new Date('%s').getTime()", x), JS)
+}
+
+parse_timeline_data <- function(.list) {
+  if (is.null(.list$group)) {
+    lapply(
+      X = seq_len(length(.list[[1]])),
+      FUN = function(i) {
+        val <- lapply(.list, `[[`, i)
+        l <- list(
+          x = as.character(val$x),
+          y = js_date(c(val$start, val$end))
+        )
+        if (!is.null(val$fill)) {
+          l$fillColor <- val$fill
+        }
+        l
+      }
+    )
+  } else {
+    grouped <- as.data.frame(.list, stringsAsFactors = FALSE)
+    grouped$group <- NULL
+    grouped <- split(
+      x = grouped,
+      f = .list$group
+    )
+    grouped <- lapply(grouped, as.list)
+    lapply(
+      X = names(grouped),
+      FUN = function(name) {
+        list(
+          name = name,
+          data = parse_timeline_data(grouped[[name]])
+        )
+      }
+    )
+  }
+}
+
+
+