+++ NEWS.md
@@ -1,3 +1,14 @@
+apexcharter 0.1.4
+==================
+
+* Upgraded ApexCharts.js to 3.17.1
+* Fixed a bug in grouped bar charts with different levels in groups.
+* New vignette to explain how to sync charts.
+* New vignette to show shiny usage.
+* Added functions `set_input_click()`, `set_input_zoom()` and `set_input_selection()` to add interaction in Shiny applications.
+
+
+
 apexcharter 0.1.3
 ==================